cloud_function/requirements.txt
google-cloud-storage
google-cloud-bigquery
functions-framework
pandas
openpyxl
pyarrow
